+++ Engine/Engine.vcxproj.filters
@@ -19,6 +19,12 @@
     <Filter Include="Input">
       <UniqueIdentifier>{81f18912-beb1-4f6d-844d-d137f4ffab74}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Render">
+      <UniqueIdentifier>{cf9feda4-801d-4aae-bc02-1167d14b7d33}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Render\Vulkan">
+      <UniqueIdentifier>{e90994da-40fa-4ba7-910e-f9547f4612ca}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Application\WinApplication.h">
@@ -57,12 +63,6 @@
     <ClInclude Include="VertexBuffer.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="VertexFormat.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="vulkanControl.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="Input\InputEventTypes.h">
       <Filter>Input</Filter>
     </ClInclude>
@@ -72,8 +72,17 @@
     <ClInclude Include="Input\InputManager.h">
       <Filter>Input</Filter>
     </ClInclude>
-    <ClInclude Include="EngineCore.h">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="Render\Renderer.h">
+      <Filter>Render</Filter>
+    </ClInclude>
+    <ClInclude Include="Render\Vulkan\VertexFormat.h">
+      <Filter>Render\Vulkan</Filter>
+    </ClInclude>
+    <ClInclude Include="Render\Vulkan\vulkanControl.h">
+      <Filter>Render\Vulkan</Filter>
+    </ClInclude>
+    <ClInclude Include="Render\Vulkan\VulkanVertexHelper.h">
+      <Filter>Render\Vulkan</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
@@ -104,14 +113,17 @@
     <ClCompile Include="stdafx.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="vulkanControl.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="Input\InputManager.cpp">
       <Filter>Input</Filter>
     </ClCompile>
-    <ClCompile Include="EngineCore.cpp">
-      <Filter>Source Files</Filter>
+    <ClCompile Include="Render\Renderer.cpp">
+      <Filter>Render</Filter>
+    </ClCompile>
+    <ClCompile Include="Render\Vulkan\vulkanControl.cpp">
+      <Filter>Render\Vulkan</Filter>
+    </ClCompile>
+    <ClCompile Include="Render\Vulkan\VulkanVertexHelper.cpp">
+      <Filter>Render\Vulkan</Filter>
     </ClCompile>
   </ItemGroup>
 </Project>
