+++ Engine/GameObject.h
@@ -1,8 +1,16 @@
 #pragma once
 
-#include "VertexFormat.h"
+#include "Render/Vulkan/VertexFormat.h"
 
-class GameObject {
+typedef struct VkBuffer_T* VkBuffer;
+typedef struct VkDeviceMemory_T* VkDeviceMemory;
+typedef struct VkDescriptorSet_T* VkDescriptorSet;
+typedef struct VkCommandBuffer_T* VkCommandBuffer;
+typedef struct VkPipeline_T* VkPipeline;
+
+
+namespace StudyEngine {
+    class STUDY_ENGINE GameObject {
     public:
         GameObject();
         ~GameObject();
@@ -15,12 +23,12 @@ class GameObject {
         void CreateIndexBuffer();
 
 
-       /* virtual void CreateUniformBuffers();*/
-       virtual void UpdateDescriptorSets();
-       virtual void BindGraphicPipeline(std::string vertShaderPath, std::string fragShaderPath);
-       virtual void CreateUniformBuffers();
+        /* virtual void CreateUniformBuffers();*/
+        virtual void UpdateDescriptorSets();
+        virtual void BindGraphicPipeline(std::string vertShaderPath, std::string fragShaderPath);
+        virtual void CreateUniformBuffers();
 
-       struct RenderData {
+        struct RenderData {
             Matrix4 m_transform;
         };
 
@@ -33,4 +41,5 @@ class GameObject {
         std::vector<Vertex> m_vertices;
         std::vector<uint32_t> m_indices;
 
-};
+    };
+}
