+++ Engine/Mesh.h
@@ -1,8 +1,8 @@
 #pragma once
 
+#include "stdafx.h"
 #include <vector>
 #include <string>
-#include <vulkan/vulkan.h>
 
 class Mesh {
     public:
