+++ Engine/Render/Renderer.h
@@ -1,0 +1,13 @@
+#pragma once
+#include "stdafx.h"
+
+namespace StudyEngine {
+    class STUDY_ENGINE Renderer {
+    public:
+        Renderer();
+        ~Renderer();
+        static bool Init();
+
+        static bool ShutDown();
+    };
+}
