+++ Engine/Render/Vulkan/VertexFormat.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "stdafx.h"
+#include "engineMath.h"
+
+struct Vertex {
+    Vector3 pos;
+    Vector3 normal;
+    Vector2 texCoord;
+};
